tests/requirements-test.txt
# This requirements are for development and testing only, not for production.
coverage
django
pytest-cov
codecov
gitchangelog